refactor: move design_creator module to nn package
