+++ elasticai/creator/nn/fixed_point/linear/layer/batch_normed_linear.py
@@ -3,12 +3,12 @@ from typing import Any, cast
 import torch
 
 from elasticai.creator.base_modules.linear import Linear as LinearBase
+from elasticai.creator.nn.design_creator import DesignCreatorModule
 from elasticai.creator.nn.fixed_point._math_operations import MathOperations
 from elasticai.creator.nn.fixed_point._two_complement_fixed_point_config import (
     FixedPointConfig,
 )
 from elasticai.creator.nn.fixed_point.linear.design import Linear as LinearDesign
-from elasticai.creator.vhdl.design_creator import DesignCreatorModule
 
 
 class BatchNormedLinear(DesignCreatorModule, torch.nn.Module):
