+++ elasticai/creator/nn/fixed_point/relu/layer.py
@@ -1,5 +1,5 @@
 from elasticai.creator.base_modules.relu import ReLU as ReLUBase
-from elasticai.creator.vhdl.design_creator import DesignCreatorModule
+from elasticai.creator.nn.design_creator import DesignCreatorModule
 
 from .design import ReLU as ReLUDesign
